fixed casting issue with atomic in connection_dispatcher_task for real this time

+++ src/network/connection_dispatcher_task.cpp
@@ -60,7 +60,7 @@ void ConnectionDispatcherTask::DispatchConnection(int fd, short) {
   }
 
   // Dispatch by rand number
-  unsigned int handler_id = next_handler_;
+  long unsigned int handler_id = next_handler_;
 
   // update next threadID
   next_handler_ = (next_handler_ + 1) % handlers_.size();
